+++ R/tycho2.R
@@ -80,7 +80,33 @@
 #'   \code{params} or \code{queryterms}.
 #' @param baseurl string. Defaults to "https://www.tycho.pitt.edu/api/".
 #'
-#' @return dataframe
+#' @return dataframe with the following possible columns:
+#'   \item{$ConditionName}{factor. Name of reported condition as listed in \href{https://doi.org/10.25504/FAIRsharing.d88s6e}{SNOMED-CT}}
+#'   \item{$ConditionSNOMED}{factor. \href{https://doi.org/10.25504/FAIRsharing.d88s6e}{SNOMED-CT} code for reported condition}
+#'   \item{$PathogenName}{factor. \href{https://doi.org/10.25504/FAIRsharing.fj07xj}{NCBI Taxonomy} organism name for pathogen causing reported condition}
+#'   \item{$PathogenTaxonID}{factor. \href{https://doi.org/10.25504/FAIRsharing.fj07xj}{NCBI Taxonomy} identifier for pathogen causing reported condition}
+#'   \item{$Fatalities}{logical. Counts of reported condition ($CountValue) represent fatalities}
+#'   \item{$CountryName}{factor. \href{https://www.iso.org/obp/ui/#search/code/}{ISO 3166} English Short Name of country}
+#'   \item{$CountryCode}{factor. \href{https://www.iso.org/obp/ui/#search/code/}{ISO 3166} 2-letter code for country}
+#'   \item{$Admin1Name}{factor. \href{https://www.iso.org/standard/63546.html}{ISO 3166-2} Name of first administrative subdivision (such as US state)}
+#'   \item{$Admin1ISO}{factor. \href{https://www.iso.org/standard/63546.html}{ISO 3166-2} code for first administrative subdivision}
+#'   \item{$Admin2Name}{factor. \href{http://www.geonames.org/}{Geonames} Placename of second order administrative division}
+#'   \item{$CityName}{factor. \href{http://www.geonames.org/}{Geonames} Name of populated place}
+#'   \item{$PeriodStartDate}{Date, format: YYYY-MM-DD. Start date of time interval for which a count was reported}
+#'   \item{$PeriodEndDate}{Date, format: YYYY-MM-DD. End date of time interval for which a count was reported}
+#'   \item{$PartOfCumulativeCountSeries}{logical. Count is part of a series of cumulative counts
+#'   (instead of being part of a series of fixed time interval counts)}
+#'   \item{$AgeRange}{Ordered factor. Age range in years for which a count was reported e.g. "0-18". Max age = 130}
+#'   \item{$Subpopulation}{factor. "Civilian", "Military", or "None specified"}
+#'   \item{$PlaceOfAcquisition}{factor. "Domestic", "Abroad", or NA}
+#'   \item{$DiagnosisCertainty}{factor. \href{https://doi.org/10.25504/FAIRsharing.d88s6e}{SNOMED-CT}
+#'   Qualifier for certainty of diagnosis for a count condition: "Definite", "Equivocal", "Possible diagnosis","Probable diagnosis", or NA}
+#'   \item{$SourceName}{factor. Name of the source (system, database, institution) from which counts
+#'   were obtained by the Project Tycho team}
+#'   \item{$CountValue}{integer. The count value.}
+#'   Variables described in detail here:
+#'   \url{https://www.tycho.pitt.edu/dataformat/ProjectTychoCustomCompiledDataFormat.pdf}
+#
 #'
 #' @examples
 #' # Note: retrive your API key from your Project Tycho account
@@ -125,5 +151,73 @@ tycho2 <- function(path="", params=NULL, queryterms=NULL, apikey=NULL, baseurl="
       out <- rbind(out,df)
     }
   }
+
+  # classes
+  vars <- colnames(out)
+
+  if("ConditionName" %in% vars) {
+    out$ConditionName <- as.factor(out$ConditionName)
+  }
+  if("ConditionSNOMED" %in% vars) {
+    out$ConditionSNOMED <- as.factor(out$ConditionSNOMED)
+  }
+  if("PathogenName" %in% vars) {
+    out$PathogenName <- as.factor(out$PathogenName)
+  }
+  if("PathogenTaxonID" %in% vars) {
+    out$PathogenTaxonID <- as.factor(out$PathogenTaxonID)
+  }
+  if("Fatalities" %in% vars) {
+    out$Fatalities <- as.logical(out$Fatalities)
+  }
+  if("CountryName" %in% vars) {
+    out$CountryName <- as.factor(out$CountryName)
+  }
+  if("CountryCode" %in% vars) {
+    out$CountryCode <- as.factor(out$CountryCode)
+  }
+  if("Admin1Name" %in% vars) {
+    out$Admin1Name <- as.factor(out$Admin1Name)
+  }
+  if("Admin1ISO" %in% vars) {
+    out$Admin1ISO <- as.factor(out$Admin1ISO)
+  }
+  if("Admin2Name" %in% vars) {
+    out$Admin2Name <- as.factor(out$Admin2Name)
+  }
+  if("CityName" %in% vars) {
+    out$CityName <- as.factor(out$CityName)
+  }
+  if("PeriodStartDate" %in% vars) {
+    out$PeriodStartDate <- as.Date(out$PeriodStartDate,format = "%Y-%m-%d")
+    }
+  if("PeriodEndDate" %in% vars) {
+    out$PeriodEndDate <- as.Date(out$PeriodEndDate,format = "%Y-%m-%d")
+  }
+  if("PartOfCumulativeCountSeries" %in% vars) {
+    out$PartOfCumulativeCountSeries <- as.logical(out$PartOfCumulativeCountSeries)
+  }
+  if("AgeRange" %in% vars) {
+    out$AgeRange <- as.ordered(out$AgeRange)
+  }
+  if("Subpopulation" %in% vars) {
+    out$Subpopulation <- factor(out$Subpopulation,
+                                levels = c("Civilian", "Military","None specified"))
+  }
+  if("PlaceOfAcquisition" %in% vars) {
+    out$PlaceOfAcquisition <- factor(out$PlaceOfAcquisition,
+                                     levels = c("Domestic", "Abroad"))
+  }
+  if("DiagnosisCertainty" %in% vars) {
+    out$DiagnosisCertainty <- as.factor(out$DiagnosisCertainty)
+  }
+  if("SourceName" %in% vars) {
+    out$SourceName <- as.factor(out$SourceName)
+  }
+  if("CountValue" %in% vars) {
+    out$CountValue <- as.integer(out$CountValue)
+  }
+
+
   return(out)
 }