+++ NEWS.md
@@ -1,3 +1,8 @@
+# tycho2 0.0.1.9000
+
+* fixed classes of returned dataframe columns
+* added documentation of returned dataframe columns
+
 # tycho2 0.0.1
 
 * Initial release